evolution/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.3
django-jenkins==0.110.0
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
